+++ src/core/main.rs
@@ -19,6 +19,7 @@ use std::collections::HashSet;
 
 use crate::core::callbacks::*;
 use crate::core::drawing_texture::*;
+use crate::core::layout_manager::*;
 use crate::core::point::*;
 use crate::core::widget_store::*;
 use crate::widget::widget::*;
@@ -60,6 +61,32 @@ impl Pushrod {
     /// Convenience method that adds a `Widget` to the GUI display stack.
     pub fn add_widget(&mut self, name: &str, widget: Box<dyn Widget>) -> i32 {
         self.widget_store.borrow_mut().add_widget(name, widget)
+    }
+
+    /// Convenience method that adds a `LayoutManager` to the layout management stack.
+    pub fn add_layout_manager(&mut self, manager: Box<dyn LayoutManager>) -> i32 {
+        self.widget_store.borrow_mut().add_layout_manager(manager)
+    }
+
+    /// Convenience method that adds a `Widget` to a `LayoutManager` by the manager's ID and
+    /// the positioning of the `Widget`.
+    pub fn add_widget_to_layout_manager(
+        &mut self,
+        name: &str,
+        widget: Box<dyn Widget>,
+        manager_id: i32,
+        position: Point,
+    ) -> i32 {
+        let widget_id = self
+            .widget_store
+            .borrow_mut()
+            .add_widget_to_layout_manager(name, widget, manager_id, position);
+
+        self.widget_store
+            .borrow_mut()
+            .do_layout_for_manager(manager_id);
+
+        widget_id
     }
 
     /// Convenience method that adds a `Widget` to a parent by its ID.  This guarantees a refresh
@@ -289,6 +316,9 @@ impl Pushrod {
 
             event.resize(|w, h| {
                 self.handle_resize(w as u32, h as u32);
+                self.widget_store
+                    .borrow_mut()
+                    .resize_layout_managers(w as u32, h as u32);
 
                 self.broadcast_event(CallbackEvent::WindowResized {
                     size: crate::core::point::Size {